* commit 'bfbe585ed4360b007640b227d49b6e8d7a2352b0':
  AArch64: Use long for pointers in audio effect classes

+++ media/java/android/media/audiofx/AudioEffect.java
@@ -315,8 +315,8 @@ public class AudioEffect {
     private int mId;
 
     // accessed by native methods
-    private int mNativeAudioEffect;
-    private int mJniData;
+    private long mNativeAudioEffect;
+    private long mJniData;
 
     /**
      * Effect descriptor
